requirements.txt
# requirements.txt
streamlit==1.24.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scipy==1.10.1
scikit-image==0.20.0
Pillow==9.5.0
imageio==2.31.1
tqdm==4.65.0
PyYAML==6.0
logging==0.5.1.2